+++ include/control.h
@@ -1,10 +1,10 @@
 // SPDX-FileCopyrightText: 2026 AnmiTaliDev <anmitalidev@nuros.org>
 // SPDX-License-Identifier: GPL-3.0-only
-// https://github.com/NurOS-Linux/neoinit
+// https://github.com/NurOS-Raesir/raesir
 
 #pragma once
 
-#define NEOINIT_SOCK_PATH "/run/neoinit.sock"
+#define RAESIR_SOCK_PATH "/run/raesir.sock"
 
 int control_init(void);
 void control_handle_data(int listen_fd);