+++ src/service.c
@@ -1,6 +1,6 @@
 // SPDX-FileCopyrightText: 2026 AnmiTaliDev <anmitalidev@nuros.org>
 // SPDX-License-Identifier: GPL-3.0-only
-// https://github.com/NurOS-Linux/neoinit
+// https://github.com/NurOS-Raesir/raesir
 
 #include "service.h"
 #include "log.h"
@@ -116,11 +116,11 @@ int service_load_dir(const char *dir, service_def_t ***out) {
 
         if (strcmp(ext, "yaml") == 0 || strcmp(ext, "yml") == 0) {
             def = service_parse_yaml(path);
-#ifdef NEOINIT_SYSTEMD_COMPAT
+#ifdef RAESIR_SYSTEMD_COMPAT
         } else if (strcmp(ext, "service") == 0) {
             def = service_parse_sd(path);
 #endif
-#ifdef NEOINIT_RUNIT_COMPAT
+#ifdef RAESIR_RUNIT_COMPAT
         } else {
             struct stat st;
             if (stat(path, &st) == 0 && S_ISDIR(st.st_mode))